refactor: split AnalysisTranscripts.tsx (1037 lines) into 7 focused modules
Split into:
- AnalysisTranscripts.tsx (289) — main page component
- AnalysisTranscriptHeader.tsx (198) — header bar
- PairedStabilityView.tsx (95) — paired stability section
- useAnalysisTranscriptsData.ts (395) — data fetching hook
- useAnalysisTranscriptParams.ts (87) — URL param parsing hook
- analysisTranscriptParams.ts (246) — pure utility functions
- analysisTranscriptFilters.ts (170) — filtering logic

Codex: 4/4 PASS. Gemini: 5/5 PASS.
