+++ cloud/apps/web/src/pages/AnalysisTranscripts.tsx
@@ -4,740 +4,127 @@
  * Shows filtered transcripts for a pivot cell in a full page view.
  */
 
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useParams, useNavigate, useSearchParams } from 'react-router-dom';
+import { useEffect, useState } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
 import { ArrowLeft } from 'lucide-react';
-import { formatTrialSignature } from '@valuerank/shared/trial-signature';
 import { Button } from '../components/ui/Button';
 import { Loading } from '../components/ui/Loading';
 import { ErrorMessage } from '../components/ui/ErrorMessage';
 import { TranscriptList } from '../components/runs/TranscriptList';
 import { TranscriptViewer } from '../components/runs/TranscriptViewer';
 import { AnalysisScopeBanner } from '../components/analysis/AnalysisScopeBanner';
-import { useRun } from '../hooks/useRun';
-import { useRuns } from '../hooks/useRuns';
-import { useAnalysis } from '../hooks/useAnalysis';
-import { useRunMutations } from '../hooks/useRunMutations';
 import type { Transcript } from '../api/operations/runs';
-import {
-  filterTranscriptsForPivotCell,
-  getScenarioDimensionsForId,
-  normalizeModelId,
-  normalizeScenarioId,
-} from '../utils/scenarioUtils';
-import { collectScenarioIdsForDecisionBucket } from '../utils/decisionBuckets';
-import {
-  deriveScenarioAttributesFromDefinition,
-  deriveDecisionDimensionLabels,
-  getDecisionSideNames,
-  mapDecisionSidesToScenarioAttributes,
-  resolveScenarioAttributes,
-  resolveScenarioAxisDimensions,
-} from '../utils/decisionLabels';
 import { formatDisplayLabel } from '../utils/displayLabels';
-import { getRunDefinitionContent } from '../utils/runDefinitionContent';
-import {
-  assertRenderableReportTranscriptSummary,
-  summarizeReportTranscriptDecisions,
-} from '../utils/reportDecisionDisplay';
-import {
-  ANALYSIS_BASE_PATH,
-  buildAnalysisTranscriptsPath,
-  isAggregateAnalysis,
-} from '../utils/analysisRouting';
-
-function getDisplaySignature(signature: string | null | undefined): string {
-  return signature && signature !== 'v?td' ? signature : 'Unknown Signature';
-}
-
-function parseConditionIds(value: string): string[] {
-  return value
-    .split(',')
-    .map((conditionId) => conditionId.trim())
-    .filter((conditionId) => conditionId.length > 0);
-}
-
-function formatRepeatPatternLabel(value: string): string {
-  switch (value) {
-    case 'stable':
-      return 'Stable';
-    case 'softLean':
-      return 'Soft Lean';
-    case 'torn':
-      return 'Torn';
-    case 'noisy':
-      return 'Unstable';
-    default:
-      return value;
-  }
-}
-
-function normalizePairedValueKey(value: string): string {
-  return value.trim().toLowerCase().replace(/\s+/g, '_');
-}
-
-type PairedConditionSource = 'current' | 'companion' | 'pooled';
-function parsePairedConditionSource(value: string | null): PairedConditionSource | null {
-  if (value === 'current' || value === 'companion' || value === 'pooled') {
-    return value;
-  }
-  return null;
-}
-
-function formatPairedConditionSourceLabel(
-  source: PairedConditionSource | null,
-  currentName?: string | null,
-  companionName?: string | null,
-): string | null {
-  if (!source) {
-    return null;
-  }
-
-  if (source === 'current') {
-    return currentName ?? 'Current vignette';
-  }
-  if (source === 'companion') {
-    return companionName ?? 'Companion vignette';
-  }
-  return 'Pooled';
-}
-
-function filterTranscriptsForConditionIds(
-  transcripts: Transcript[],
-  modelId: string,
-  conditionIds: string[],
-  scenarioDimensions: Record<string, Record<string, string | number>> | null | undefined,
-  rowDim: string,
-  colDim: string,
-): Transcript[] {
-  if (!modelId) {
-    return [];
-  }
-
-  const selectedModelNormalized = normalizeModelId(modelId);
-  const canonicalIds = new Set(conditionIds);
-  const normalizedIds = new Set(conditionIds.map((conditionId) => normalizeScenarioId(conditionId)));
-
-  return transcripts.filter((transcript) => {
-    const transcriptModelNormalized = normalizeModelId(transcript.modelId);
-    const modelMatches = transcript.modelId === modelId
-      || transcriptModelNormalized === selectedModelNormalized
-      || transcript.modelId.includes(modelId)
-      || modelId.includes(transcript.modelId);
-
-    if (!modelMatches) {
-      return false;
-    }
-    if (!transcript.scenarioId) {
-      return false;
-    }
-
-    const transcriptScenarioId = String(transcript.scenarioId);
-    const transcriptScenarioNormalized = normalizeScenarioId(transcriptScenarioId);
-
-    if (canonicalIds.has(transcriptScenarioId) || normalizedIds.has(transcriptScenarioNormalized)) {
-      return true;
-    }
-
-    if (!scenarioDimensions || !rowDim || !colDim) {
-      return false;
-    }
-
-    const dimensions = getScenarioDimensionsForId(transcript.scenarioId, scenarioDimensions);
-    if (!dimensions) {
-      return false;
-    }
-
-    const conditionId = `${String(dimensions[rowDim] ?? 'N/A')}||${String(dimensions[colDim] ?? 'N/A')}`;
-    return canonicalIds.has(conditionId);
-  });
-}
+import { formatPairedConditionSourceLabel } from '../utils/analysisTranscriptParams';
+import { useAnalysisTranscriptsData } from '../hooks/useAnalysisTranscriptsData';
+import { AnalysisTranscriptHeader } from './AnalysisTranscriptHeader';
+import { PairedStabilityView } from './PairedStabilityView';
 
 export function AnalysisTranscripts() {
   const navigate = useNavigate();
   const { id } = useParams<{ id: string }>();
-  const [searchParams, setSearchParams] = useSearchParams();
   const [selectedTranscript, setSelectedTranscript] = useState<Transcript | null>(null);
-  const [updatingTranscriptIds, setUpdatingTranscriptIds] = useState<Set<string>>(new Set());
 
-  const rowDim = searchParams.get('rowDim') ?? '';
-  const colDim = searchParams.get('colDim') ?? '';
-  const row = searchParams.get('row') ?? '';
-  const col = searchParams.get('col') ?? '';
-  const selectedModel = searchParams.get('modelId') ?? searchParams.get('model') ?? '';
-  const favoredValueKey = searchParams.get('favoredValueKey') ?? '';
-  const decisionStrength = searchParams.get('decisionStrength') ?? '';
-  const normalizedFavoredValueKey = favoredValueKey === '' ? undefined : favoredValueKey;
-  const normalizedDecisionStrength = decisionStrength === 'strong'
-    || decisionStrength === 'lean'
-    || decisionStrength === 'neutral'
-    || decisionStrength === 'unknown'
-    ? decisionStrength
-    : undefined;
-  const decisionBucket = searchParams.get('decisionBucket') ?? '';
-  const repeatPattern = searchParams.get('repeatPattern') ?? '';
-  const selectedTranscriptId = searchParams.get('transcriptId') ?? '';
-  const companionRunId = searchParams.get('companionRunId') ?? '';
-  const pairedValueKey = searchParams.get('pairedValueKey') ?? '';
-  const pairedDecisionBucketParam = searchParams.get('pairedDecisionBucket') ?? '';
-  const pairedValueLabel = searchParams.get('pairedValueLabel') ?? '';
-  const pairView = searchParams.get('pairView') ?? '';
-  const orientationBucketParam = searchParams.get('orientationBucket');
-  const hasLegacyOrientationBucket = orientationBucketParam === 'canonical' || orientationBucketParam === 'flipped';
-  const pairedConditionSource = parsePairedConditionSource(searchParams.get('sourceRun'));
-  const analysisMode = searchParams.get('mode') === 'paired'
-    ? 'paired'
-    : searchParams.get('mode') === 'single'
-      ? 'single'
-      : null;
-  const conditionIds = useMemo(
-    () => parseConditionIds(searchParams.get('conditionIds') ?? ''),
-    [searchParams]
-  );
-  const primaryConditionIds = useMemo(
-    () => parseConditionIds(searchParams.get('primaryConditionIds') ?? ''),
-    [searchParams]
-  );
-  const companionConditionIds = useMemo(
-    () => parseConditionIds(searchParams.get('companionConditionIds') ?? ''),
-    [searchParams]
-  );
-  const isPairedStabilityDrilldown = Boolean(
-    repeatPattern && companionRunId && searchParams.has('primaryConditionIds')
-  );
-
-  const { run, loading, error, refetch } = useRun({
-    id: id || '',
-    pause: !id,
-    enablePolling: false,
-  });
-
-  const { updateTranscriptDecision } = useRunMutations();
-
-  const { analysis } = useAnalysis({
-    runId: id || '',
-    pause: !id,
-    enablePolling: false,
-    analysisStatus: run?.analysisStatus ?? null,
-  });
-  const { run: companionRun } = useRun({
-    id: companionRunId,
-    pause: companionRunId === '',
-    enablePolling: false,
-  });
-  const { analysis: companionAnalysis } = useAnalysis({
-    runId: companionRunId,
-    pause: companionRunId === '' || !companionRun?.analysisStatus,
-    enablePolling: false,
-    analysisStatus: companionRun?.analysisStatus ?? null,
-  });
-
-  const isAggregate = isAggregateAnalysis(
-    run?.tags?.some((tag) => tag.name === 'Aggregate') ?? false,
-    analysis?.analysisType,
-  );
-
-  const config = run?.config as {
-    definitionSnapshot?: { _meta?: { definitionVersion?: unknown }, version?: unknown };
-    temperature?: unknown;
-  } | null;
-
-  const runDefinitionVersion = typeof config?.definitionSnapshot?._meta?.definitionVersion === 'number'
-    ? config.definitionSnapshot._meta.definitionVersion
-    : typeof config?.definitionSnapshot?.version === 'number'
-      ? config.definitionSnapshot.version
-      : typeof run?.definitionVersion === 'number'
-        ? run.definitionVersion
-        : null;
-
-  const runTemperature = typeof config?.temperature === 'number' ? config.temperature : null;
-  const trialSignature = formatTrialSignature(runDefinitionVersion, runTemperature);
-
-  const { runs } = useRuns({
-    definitionId: isAggregate ? (run?.definition?.id || undefined) : undefined,
-    status: 'COMPLETED',
-    limit: 1000,
-    pause: !isAggregate || !run?.definition?.id,
-  });
-
-  const aggregateRuns = useMemo(() => {
-    return runs
-      .filter((candidate) => candidate.tags.some((tag) => tag.name === 'Aggregate'))
-      .map((candidate) => {
-        const candidateConfig = candidate.config as {
-          definitionSnapshot?: { _meta?: { definitionVersion?: unknown }, version?: unknown };
-          temperature?: unknown;
-        } | null;
-        const candidateVersion = typeof candidateConfig?.definitionSnapshot?._meta?.definitionVersion === 'number'
-          ? candidateConfig.definitionSnapshot._meta.definitionVersion
-          : typeof candidateConfig?.definitionSnapshot?.version === 'number'
-            ? candidateConfig.definitionSnapshot.version
-            : typeof candidate.definitionVersion === 'number'
-              ? candidate.definitionVersion
-              : null;
-        const candidateTemp = typeof candidateConfig?.temperature === 'number'
-          ? candidateConfig.temperature
-          : null;
-
-        return {
-          id: candidate.id,
-          signature: formatTrialSignature(candidateVersion, candidateTemp),
-        };
-      })
-      .reduce<Array<{ id: string; signature: string; count: number }>>((acc, candidate) => {
-        const existing = acc.find((item) => item.signature === candidate.signature);
-        if (existing) {
-          existing.count += 1;
-          if (candidate.id === run?.id) {
-            existing.id = candidate.id;
-          }
-          return acc;
-        }
-        acc.push({ ...candidate, count: 1 });
-        return acc;
-      }, []);
-  }, [run?.id, runs]);
-
-  const selectedAggregateSignature = aggregateRuns.find((candidate) => candidate.id === run?.id)?.signature
-    ?? trialSignature;
-
-  const scenarioDimensions = analysis?.visualizationData?.scenarioDimensions;
-  const modelScenarioMatrix = analysis?.visualizationData?.modelScenarioMatrix;
-  const companionScenarioDimensions = companionAnalysis?.visualizationData?.scenarioDimensions;
-  const companionModelScenarioMatrix = companionAnalysis?.visualizationData?.modelScenarioMatrix;
-  const definitionContent = useMemo(() => getRunDefinitionContent(run), [run]);
-  const companionDefinitionContent = useMemo(() => getRunDefinitionContent(companionRun), [companionRun]);
-  const preferredAttributes = useMemo(
-    () => deriveScenarioAttributesFromDefinition(definitionContent),
-    [definitionContent]
-  );
-  const availableAttributes = useMemo(() => {
-    return resolveScenarioAttributes(scenarioDimensions, preferredAttributes, modelScenarioMatrix);
-  }, [scenarioDimensions, preferredAttributes, modelScenarioMatrix]);
-  const mergedScenarioDimensions = useMemo(
-    () => ({
-      ...(scenarioDimensions ?? {}),
-      ...(companionScenarioDimensions ?? {}),
-    }),
-    [scenarioDimensions, companionScenarioDimensions],
-  );
-  const resolvedAxes = useMemo(
-    () => resolveScenarioAxisDimensions(availableAttributes, rowDim, colDim),
-    [availableAttributes, colDim, rowDim]
-  );
-  const activeRowDim = resolvedAxes.rowDim;
-  const activeColDim = resolvedAxes.colDim;
-  const hasCellFilterParams = Boolean(activeRowDim && activeColDim && row && col);
-  const hasRepeatPatternParams = Boolean(selectedModel && repeatPattern && searchParams.has('conditionIds'));
-  const hasDirectTranscriptParam = selectedTranscriptId.length > 0;
-  const hasBucketFilterParams = Boolean(
-    activeRowDim
-    && activeColDim
-    && selectedModel
-    && (decisionBucket === 'a' || decisionBucket === 'neutral' || decisionBucket === 'b')
-  );
-  const hasPairedValueFilterParams = Boolean(
-    analysisMode === 'paired'
-    && companionRunId
-    && selectedModel
-    && (pairedValueKey || pairedDecisionBucketParam === 'a' || pairedDecisionBucketParam === 'neutral' || pairedDecisionBucketParam === 'b')
-    && pairView === 'blended'
-  );
-  const hasPairedConditionFilterParams = Boolean(
-    analysisMode === 'paired'
-    && companionRunId
-    && selectedModel
-    && activeRowDim
-    && activeColDim
-    && row
-    && col
-    && (pairView === 'condition-blended' || pairView === 'condition-split')
-  );
-  const pairedConditionStateError = useMemo(() => {
-    if (hasLegacyOrientationBucket && analysisMode === 'paired') {
-      return new Error('Legacy orientationBucket URLs are no longer supported. Use sourceRun=current, sourceRun=companion, or sourceRun=pooled.');
-    }
-
-    if (!hasPairedConditionFilterParams) {
-      return null;
-    }
-
-    if (pairView !== 'condition-split') {
-      return null;
-    }
-
-    if (pairedConditionSource === 'current' || pairedConditionSource === 'companion') {
-      return null;
-    }
-
-    return new Error('Split paired condition inspection requires sourceRun=current or sourceRun=companion.');
-  }, [analysisMode, hasLegacyOrientationBucket, hasPairedConditionFilterParams, pairView, pairedConditionSource]);
-  const dimensionLabels = useMemo(
-    () => deriveDecisionDimensionLabels(definitionContent),
-    [definitionContent]
-  );
-
-  const decisionSideNames = useMemo(
-    () => getDecisionSideNames(dimensionLabels),
-    [dimensionLabels]
-  );
-  const bucketAttributes = useMemo(
-    () => mapDecisionSidesToScenarioAttributes(decisionSideNames.aName, decisionSideNames.bName, [activeRowDim, activeColDim].filter((d) => d !== '')),
-    [activeColDim, activeRowDim, decisionSideNames.aName, decisionSideNames.bName]
-  );
-
-  const decisionBucketLabel = useMemo(() => {
-    if (decisionBucket === 'a') return bucketAttributes.lowAttribute;
-    if (decisionBucket === 'b') return bucketAttributes.highAttribute;
-    if (decisionBucket === 'neutral') return 'Neutral';
-    return '';
-  }, [bucketAttributes.highAttribute, bucketAttributes.lowAttribute, decisionBucket]);
-
-  const resolveDecisionBucketForValue = useCallback((
-    content: unknown,
-    rowDimension: string,
-    colDimension: string,
-    targetValueKey: string,
-  ): 'a' | 'b' | null => {
-    const labels = deriveDecisionDimensionLabels(content);
-    const sideNames = getDecisionSideNames(labels);
-    const mapped = mapDecisionSidesToScenarioAttributes(
-      sideNames.aName,
-      sideNames.bName,
-      [rowDimension, colDimension].filter((value) => value !== ''),
-    );
-    const normalizedTarget = normalizePairedValueKey(targetValueKey);
-    if (normalizePairedValueKey(mapped.lowAttribute) === normalizedTarget) return 'a';
-    if (normalizePairedValueKey(mapped.highAttribute) === normalizedTarget) return 'b';
-    return null;
-  }, []);
-
-  useEffect(() => {
-    if (!scenarioDimensions) return;
-    if (hasRepeatPatternParams) return;
-    if (isPairedStabilityDrilldown) return;
-
-    const rowChanged = rowDim !== activeRowDim;
-    const colChanged = colDim !== activeColDim;
-    if (!rowChanged && !colChanged) return;
-
-    const nextParams = new URLSearchParams(searchParams);
-    nextParams.set('rowDim', activeRowDim);
-    nextParams.set('colDim', activeColDim);
-
-    // Existing row/col cell coordinates may not be valid once axes change.
-    if (rowChanged) nextParams.delete('row');
-    if (colChanged) nextParams.delete('col');
-
-    setSearchParams(nextParams, { replace: true });
-  }, [
-    scenarioDimensions,
+  const {
+    loading,
+    error,
+    run,
+    companionRun,
+    isAggregate,
+    trialSignature,
+    selectedAggregateSignature,
+    aggregateRuns,
     searchParams,
-    setSearchParams,
-    rowDim,
-    colDim,
-    activeRowDim,
-    activeColDim,
-    hasRepeatPatternParams,
-    isPairedStabilityDrilldown,
-  ]);
-
-  const handleDecisionChange = useCallback(async (transcript: Transcript, nextDecisionCode: string) => {
-    setUpdatingTranscriptIds((prev) => new Set(prev).add(transcript.id));
-    try {
-      await updateTranscriptDecision(transcript.id, nextDecisionCode);
-      refetch();
-    } finally {
-      setUpdatingTranscriptIds((prev) => {
-        const next = new Set(prev);
-        next.delete(transcript.id);
-        return next;
-      });
-    }
-  }, [updateTranscriptDecision, refetch]);
-
-  const filteredTranscripts = useMemo(() => {
-    if (hasDirectTranscriptParam) {
-      const matched = (run?.transcripts ?? []).find((transcript) => transcript.id === selectedTranscriptId);
-      return matched ? [matched] : [];
-    }
-
-    if (hasRepeatPatternParams) {
-      return filterTranscriptsForConditionIds(
-        run?.transcripts ?? [],
-        selectedModel,
-        conditionIds,
-        scenarioDimensions,
-        activeRowDim,
-        activeColDim,
-      );
-    }
-
-    if (hasPairedConditionFilterParams) {
-      const runEntries = [
-        {
-          run,
-          source: 'current' as const,
-          content: definitionContent,
-          scenarioDims: scenarioDimensions,
-        },
-        {
-          run: companionRun,
-          source: 'companion' as const,
-          content: companionDefinitionContent,
-          scenarioDims: companionScenarioDimensions,
-        },
-      ];
-
-      return runEntries.flatMap((entry) => {
-        if (!entry.run) {
-          return [];
-        }
-        if (pairedConditionSource === 'current' && entry.source !== 'current') {
-          return [];
-        }
-        if (pairedConditionSource === 'companion' && entry.source !== 'companion') {
-          return [];
-        }
-        if (
-          pairedConditionSource === 'current'
-          || pairedConditionSource === 'companion'
-          || pairedConditionSource === 'pooled'
-        ) {
-          return filterTranscriptsForPivotCell({
-            transcripts: entry.run.transcripts ?? [],
-            scenarioDimensions: entry.scenarioDims,
-            rowDim: activeRowDim,
-            colDim: activeColDim,
-            row,
-            col,
-            selectedModel,
-            favoredValueKey: normalizedFavoredValueKey,
-            decisionStrength: normalizedDecisionStrength,
-          });
-        }
-
-        return filterTranscriptsForPivotCell({
-          transcripts: entry.run.transcripts ?? [],
-          scenarioDimensions: entry.scenarioDims,
-          rowDim: activeRowDim,
-          colDim: activeColDim,
-          row,
-          col,
-          selectedModel,
-          favoredValueKey: normalizedFavoredValueKey,
-          decisionStrength: normalizedDecisionStrength,
-        });
-      });
-    }
-
-    if (hasPairedValueFilterParams) {
-      const runEntries = [
-        {
-          run,
-          content: definitionContent,
-          scenarioDims: scenarioDimensions,
-          modelMatrix: modelScenarioMatrix,
-        },
-        {
-          run: companionRun,
-          content: companionDefinitionContent,
-          scenarioDims: companionScenarioDimensions,
-          modelMatrix: companionModelScenarioMatrix,
-        },
-      ];
-
-      return runEntries.flatMap((entry) => {
-        const transcripts = entry.run?.transcripts ?? [];
-        const attributes = deriveScenarioAttributesFromDefinition(entry.content);
-        const rowDimension = attributes[0] ?? '';
-        const colDimension = attributes[1] ?? attributes[0] ?? '';
-        const bucket = pairedDecisionBucketParam === 'a' || pairedDecisionBucketParam === 'neutral' || pairedDecisionBucketParam === 'b'
-          ? pairedDecisionBucketParam
-          : resolveDecisionBucketForValue(entry.content, rowDimension, colDimension, pairedValueKey);
-        if (!bucket) {
-          return [];
-        }
-
-        const matchingScenarioIds = collectScenarioIdsForDecisionBucket(
-          entry.scenarioDims,
-          entry.modelMatrix,
-          selectedModel,
-          bucket,
-          rowDimension,
-          colDimension,
-        );
-
-        return transcripts.filter((transcript) => (
-          transcript.modelId === selectedModel
-          && transcript.scenarioId !== null
-          && matchingScenarioIds.has(transcript.scenarioId)
-        ));
-      });
-    }
-
-    if (hasBucketFilterParams) {
-      const transcripts = run?.transcripts ?? [];
-      if (!scenarioDimensions || !modelScenarioMatrix) return [];
-      const matchingScenarioIds = collectScenarioIdsForDecisionBucket(
-        scenarioDimensions,
-        modelScenarioMatrix,
-        selectedModel,
-        decisionBucket as 'a' | 'neutral' | 'b',
-        activeRowDim,
-        activeColDim,
-      );
-
-        return transcripts.filter((transcript) => (
-          transcript.modelId === selectedModel
-          && transcript.scenarioId !== null
-          && matchingScenarioIds.has(transcript.scenarioId)
-        ));
-      }
-
-    return filterTranscriptsForPivotCell({
-      transcripts: run?.transcripts ?? [],
-      scenarioDimensions,
-      rowDim: activeRowDim,
-      colDim: activeColDim,
-      row,
-      col,
-      selectedModel,
-      favoredValueKey: normalizedFavoredValueKey,
-      decisionStrength: normalizedDecisionStrength,
-    });
-  }, [
-    scenarioDimensions,
-    modelScenarioMatrix,
-    activeRowDim,
-    activeColDim,
     row,
     col,
     selectedModel,
-    normalizedFavoredValueKey,
-    normalizedDecisionStrength,
-    conditionIds,
-    decisionBucket,
-    hasRepeatPatternParams,
-    hasPairedConditionFilterParams,
-    hasPairedValueFilterParams,
-    hasBucketFilterParams,
-    hasDirectTranscriptParam,
-    pairedConditionSource,
+    repeatPattern,
     selectedTranscriptId,
-    companionRun,
-    run,
-    companionDefinitionContent,
+    pairedValueLabel,
+    pairView,
+    pairedConditionSource,
+    analysisMode,
+    conditionIds,
+    activeRowDim,
+    activeColDim,
+    hasCellFilterParams,
+    hasRepeatPatternParams,
+    hasDirectTranscriptParam,
+    hasBucketFilterParams,
+    hasPairedValueFilterParams,
+    hasPairedConditionFilterParams,
+    isPairedStabilityDrilldown,
+    scenarioDimensions,
     companionScenarioDimensions,
-    companionModelScenarioMatrix,
-    definitionContent,
-    pairedValueKey,
-    pairedDecisionBucketParam,
-    resolveDecisionBucketForValue,
-  ]);
+    mergedScenarioDimensions,
+    decisionBucketLabel,
+    filteredTranscripts,
+    primaryStabilityTranscripts,
+    companionStabilityTranscripts,
+    decisionSummary,
+    pairedConditionStateError,
+    reportStateError,
+    handleDecisionChange,
+    updatingTranscriptIds,
+  } = useAnalysisTranscriptsData(id);
 
-  const primaryStabilityTranscripts = useMemo(() => {
-    if (!isPairedStabilityDrilldown) return [];
-    return filterTranscriptsForConditionIds(
-      run?.transcripts ?? [],
-      selectedModel,
-      primaryConditionIds,
-      scenarioDimensions,
-      activeRowDim,
-      activeColDim,
-    );
-  }, [isPairedStabilityDrilldown, run, selectedModel, primaryConditionIds, scenarioDimensions, activeRowDim, activeColDim]);
-
-  const companionStabilityTranscripts = useMemo(() => {
-    if (!isPairedStabilityDrilldown) return [];
-    return filterTranscriptsForConditionIds(
-      companionRun?.transcripts ?? [],
-      selectedModel,
-      companionConditionIds,
-      companionScenarioDimensions,
-      activeRowDim,
-      activeColDim,
-    );
-  }, [isPairedStabilityDrilldown, companionRun, selectedModel, companionConditionIds, companionScenarioDimensions, activeRowDim, activeColDim]);
-
-  const decisionSummary = useMemo(
-    () => summarizeReportTranscriptDecisions(filteredTranscripts),
-    [filteredTranscripts],
-  );
-  const reportStateError = useMemo(() => {
-    try {
-      assertRenderableReportTranscriptSummary(decisionSummary);
-      return null;
-    } catch (error) {
-      return error instanceof Error
-        ? error
-        : new Error('AnalysisTranscripts requires canonical decisionModelV2 data for every visible transcript.');
-    }
-  }, [decisionSummary]);
-  const listDisplayMode = 'audit' as const;
-  const viewerDisplayMode = 'audit' as const;
-  const decisionColumnLabel = 'Decision summary';
-  const decisionColumnTooltip = 'Shows the canonical decision headline and summary from the backend transcript data.';
   useEffect(() => {
-    if (!hasDirectTranscriptParam) {
-      return;
-    }
-    const matched = filteredTranscripts.find((transcript) => transcript.id === selectedTranscriptId) ?? null;
+    if (!hasDirectTranscriptParam) return;
+    const matched = filteredTranscripts.find((t) => t.id === selectedTranscriptId) ?? null;
     setSelectedTranscript(matched);
   }, [filteredTranscripts, hasDirectTranscriptParam, selectedTranscriptId]);
 
   useEffect(() => {
-    if (!selectedTranscript) {
-      return;
-    }
-
-    const matched = filteredTranscripts.find((transcript) => transcript.id === selectedTranscript.id);
-    if (!matched || matched === selectedTranscript) {
-      return;
-    }
-
+    if (!selectedTranscript) return;
+    const matched = filteredTranscripts.find((t) => t.id === selectedTranscript.id);
+    if (!matched || matched === selectedTranscript) return;
     setSelectedTranscript(matched);
   }, [filteredTranscripts, selectedTranscript]);
 
-  const transcriptCoverage = useMemo(() => {
-    return filteredTranscripts.reduce(
-      (acc, transcript) => {
-        const metadata = transcript.decisionModelV2?.raw;
-        const canonical = transcript.decisionModelV2?.canonical ?? null;
-        if (metadata?.manualOverride) {
-          acc.manual += 1;
-        }
-        if (metadata?.parseClass === 'exact') {
-          acc.exact += 1;
-        } else if (metadata?.parseClass === 'fallback_resolved') {
-          acc.fallback += 1;
-        } else if (metadata?.parseClass === 'ambiguous') {
-          acc.ambiguous += 1;
-        }
-        if (
-          canonical == null
-          || canonical.direction === 'unknown'
-          || canonical.strength === 'unknown'
-          || canonical.source === 'error'
-          || metadata?.parseClass === 'unparseable'
-        ) {
-          acc.unresolved += 1;
-        }
-        return acc;
-      },
-      { exact: 0, fallback: 0, ambiguous: 0, manual: 0, unresolved: 0 }
-    );
-  }, [filteredTranscripts]);
+  const transcriptCoverage = filteredTranscripts.reduce(
+    (acc, transcript) => {
+      const metadata = transcript.decisionModelV2?.raw;
+      const canonical = transcript.decisionModelV2?.canonical ?? null;
+      if (metadata?.manualOverride) acc.manual += 1;
+      if (metadata?.parseClass === 'exact') {
+        acc.exact += 1;
+      } else if (metadata?.parseClass === 'fallback_resolved') {
+        acc.fallback += 1;
+      } else if (metadata?.parseClass === 'ambiguous') {
+        acc.ambiguous += 1;
+      }
+      if (
+        canonical == null
+        || canonical.direction === 'unknown'
+        || canonical.strength === 'unknown'
+        || canonical.source === 'error'
+        || metadata?.parseClass === 'unparseable'
+      ) {
+        acc.unresolved += 1;
+      }
+      return acc;
+    },
+    { exact: 0, fallback: 0, ambiguous: 0, manual: 0, unresolved: 0 }
+  );
+
+  const listDisplayMode = 'audit' as const;
+  const viewerDisplayMode = 'audit' as const;
+  const decisionColumnLabel = 'Decision summary';
+  const decisionColumnTooltip = 'Shows the canonical decision headline and summary from the backend transcript data.';
+
+  const backButton = (
+    <div className="flex items-center gap-4">
+      <Button variant="ghost" size="sm" onClick={() => navigate(-1)}>
+        <ArrowLeft className="w-4 h-4 mr-1" />
+        Back
+      </Button>
+    </div>
+  );
 
   if (loading && !run) {
     return (
       <div className="space-y-6">
-        <div className="flex items-center gap-4">
-          <Button variant="ghost" size="sm" onClick={() => navigate(-1)}>
-            <ArrowLeft className="w-4 h-4 mr-1" />
-            Back
-          </Button>
-        </div>
+        {backButton}
         <Loading size="lg" text="Loading transcripts..." />
       </div>
     );
@@ -746,12 +133,7 @@ export function AnalysisTranscripts() {
   if (error) {
     return (
       <div className="space-y-6">
-        <div className="flex items-center gap-4">
-          <Button variant="ghost" size="sm" onClick={() => navigate(-1)}>
-            <ArrowLeft className="w-4 h-4 mr-1" />
-            Back
-          </Button>
-        </div>
+        {backButton}
         <ErrorMessage message={`Failed to load transcripts: ${error.message}`} />
       </div>
     );
@@ -760,12 +142,7 @@ export function AnalysisTranscripts() {
   if (!run) {
     return (
       <div className="space-y-6">
-        <div className="flex items-center gap-4">
-          <Button variant="ghost" size="sm" onClick={() => navigate(-1)}>
-            <ArrowLeft className="w-4 h-4 mr-1" />
-            Back
-          </Button>
-        </div>
+        {backButton}
         <ErrorMessage message="Trial not found" />
       </div>
     );
@@ -773,120 +150,32 @@ export function AnalysisTranscripts() {
 
   return (
     <div className="space-y-6">
-      <div className="flex items-center gap-4">
-        <Button variant="ghost" size="sm" onClick={() => navigate(-1)}>
-          <ArrowLeft className="w-4 h-4 mr-1" />
-          Back
-        </Button>
-        <div className="text-sm text-gray-500 flex flex-wrap items-center gap-2">
-          <span className="text-gray-700">{run.definition?.name || 'Unnamed Definition'}</span>
-          <span className="text-gray-300">•</span>
-          {isAggregate ? (
-            <>
-              <span className="bg-indigo-100 text-indigo-800 text-xs px-2 py-0.5 rounded-full font-medium">
-                Aggregate View
-              </span>
-              <span className="font-mono bg-gray-100 text-gray-800 text-xs px-2 py-0.5 rounded border border-gray-200">
-                {aggregateRuns.length > 1 ? (
-                  <select
-                    className="bg-transparent border-none p-0 pr-4 text-xs font-mono cursor-pointer focus:ring-0 focus:outline-none focus:bg-gray-200"
-                    value={selectedAggregateSignature}
-                    onChange={(e) => {
-                      const nextRun = aggregateRuns.find((candidate) => candidate.signature === e.target.value);
-                      if (nextRun) {
-                        navigate(buildAnalysisTranscriptsPath(ANALYSIS_BASE_PATH, nextRun.id, searchParams));
-                      }
-                    }}
-                  >
-                    {aggregateRuns.map((candidate) => (
-                      <option key={candidate.signature} value={candidate.signature}>
-                        {getDisplaySignature(candidate.signature)}{candidate.count > 1 ? ` (${candidate.count} runs)` : ''}
-                      </option>
-                    ))}
-                  </select>
-                ) : (
-                  getDisplaySignature(selectedAggregateSignature)
-                )}
-              </span>
-            </>
-          ) : (
-            <>
-              <span className="font-mono">Trial {run.id.slice(0, 8)}...</span>
-              <span className="font-mono bg-gray-100 text-gray-800 text-xs px-2 py-0.5 rounded border border-gray-200">
-                {getDisplaySignature(trialSignature)}
-              </span>
-            </>
-          )}
-          <span className="text-gray-300">•</span>
-          <div className="contents">
-              {isPairedStabilityDrilldown ? (
-                <>
-                  Repeat Pattern: <span className="font-medium text-gray-900">{formatRepeatPatternLabel(repeatPattern)}</span>
-                  <span className="mx-2">•</span>
-                  Model: <span className="font-medium text-gray-900">{selectedModel}</span>
-                  <span className="mx-2">•</span>
-                  <span className="font-medium text-gray-900">Both vignette orders</span>
-                </>
-              ) : hasRepeatPatternParams ? (
-                <>
-                  Repeat Pattern: <span className="font-medium text-gray-900">{formatRepeatPatternLabel(repeatPattern)}</span>
-                  <span className="mx-2">•</span>
-                  Model: <span className="font-medium text-gray-900">{selectedModel}</span>
-                  <span className="mx-2">•</span>
-                  Conditions: <span className="font-medium text-gray-900">{conditionIds.length}</span>
-                  {decisionSummary.headline !== '—' && (
-                    <>
-                      <span className="mx-2">•</span>
-                      Decision summary: <span className="font-medium text-gray-900">{decisionSummary.headline}</span>
-                    </>
-                  )}
-                </>
-              ) : (activeRowDim && activeColDim) ? (
-                <>
-                  {hasCellFilterParams ? (
-                    <>
-                      {activeRowDim}: <span className="font-medium text-gray-900">{row || '-'}</span>
-                      <span className="mx-2">•</span>
-                      {activeColDim}: <span className="font-medium text-gray-900">{col || '-'}</span>
-                    </>
-                  ) : (
-                    <>
-                      {activeRowDim} × {activeColDim}
-                    </>
-                  )}
-              <span className="mx-2">•</span>
-              Model: <span className="font-medium text-gray-900">{selectedModel || 'All Models'}</span>
-              {decisionBucketLabel && (
-                <>
-                  <span className="mx-2">•</span>
-                  Favors: <span className="font-medium text-gray-900">{decisionBucketLabel}</span>
-                </>
-              )}
-              {pairedConditionSource && (
-                <>
-                  <span className="mx-2">•</span>
-                  Source: <span className="font-medium text-gray-900">{formatPairedConditionSourceLabel(pairedConditionSource, run?.definition?.name, companionRun?.definition?.name)}</span>
-                </>
-              )}
-              {pairedValueLabel && (
-                <>
-                  <span className="mx-2">•</span>
-                  Paired Value: <span className="font-medium text-gray-900">{formatDisplayLabel(pairedValueLabel)}</span>
-                </>
-              )}
-              {decisionSummary.headline !== '—' && (
-                <>
-                  <span className="mx-2">•</span>
-                  Decision summary: <span className="font-medium text-gray-900">{decisionSummary.headline}</span>
-                </>
-              )}
-            </>
-          ) : (
-            'Transcript Filter'
-          )}
-          </div>
-        </div>
-      </div>
+      <AnalysisTranscriptHeader
+        onBack={() => navigate(-1)}
+        onNavigate={(path) => navigate(path)}
+        runId={run.id}
+        definitionName={run.definition?.name}
+        isAggregate={isAggregate}
+        trialSignature={trialSignature}
+        selectedAggregateSignature={selectedAggregateSignature}
+        aggregateRuns={aggregateRuns}
+        searchParams={searchParams}
+        activeRowDim={activeRowDim}
+        activeColDim={activeColDim}
+        row={row}
+        col={col}
+        selectedModel={selectedModel}
+        decisionBucketLabel={decisionBucketLabel}
+        pairedConditionSource={pairedConditionSource}
+        companionDefinitionName={companionRun?.definition?.name}
+        pairedValueLabel={pairedValueLabel}
+        decisionSummary={decisionSummary}
+        repeatPattern={repeatPattern}
+        conditionIds={conditionIds}
+        isPairedStabilityDrilldown={isPairedStabilityDrilldown}
+        hasRepeatPatternParams={hasRepeatPatternParams}
+        hasCellFilterParams={hasCellFilterParams}
+      />
 
       {analysisMode && (
         <AnalysisScopeBanner analysisMode={analysisMode} compact />
@@ -921,9 +210,7 @@ export function AnalysisTranscripts() {
       {!isPairedStabilityDrilldown && filteredTranscripts.length > 0 && (
         <div className="rounded-lg border border-gray-200 bg-white p-4 text-sm text-gray-700">
           <div className="flex flex-wrap items-center gap-2">
-            <span className="font-medium text-gray-900">
-              Decision coverage
-            </span>
+            <span className="font-medium text-gray-900">Decision coverage</span>
             <span className="rounded-full bg-gray-100 px-2 py-0.5 text-xs text-gray-700">
               Exact {transcriptCoverage.exact}
             </span>
@@ -947,56 +234,21 @@ export function AnalysisTranscripts() {
       )}
 
       {isPairedStabilityDrilldown ? (
-        <div className="space-y-6">
-          <div className="rounded-lg border border-teal-200 bg-teal-50 p-3 text-sm text-teal-800">
-            Stability drilldown is active for <span className="font-medium">{formatRepeatPatternLabel(repeatPattern)}</span>. Transcripts are shown separately for each vignette order.
-          </div>
-          {primaryStabilityTranscripts.length === 0 && companionStabilityTranscripts.length === 0 && (
-            <div className="rounded-lg border border-gray-200 bg-white p-6 text-sm text-gray-500">
-              No transcripts match the selected stability pattern.
-            </div>
-          )}
-          <section className="space-y-3">
-            <h3 className="text-sm font-semibold text-gray-700">{run?.definition?.name ?? 'Primary vignette order'}</h3>
-            {primaryStabilityTranscripts.length > 0 ? (
-              <TranscriptList
-                transcripts={primaryStabilityTranscripts}
-                onSelect={setSelectedTranscript}
-                groupByModel={false}
-                scenarioDimensions={scenarioDimensions}
-                onDecisionChange={handleDecisionChange}
-                updatingTranscriptIds={updatingTranscriptIds}
-                decisionColumnLabel={decisionColumnLabel}
-                decisionColumnTooltip={decisionColumnTooltip}
-                decisionDisplayMode={listDisplayMode}
-              />
-            ) : (
-              <div className="rounded-lg border border-gray-200 bg-white p-4 text-sm text-gray-500">
-                No transcripts matched this pattern for this vignette order.
-              </div>
-            )}
-          </section>
-          <section className="space-y-3">
-            <h3 className="text-sm font-semibold text-gray-700">{companionRun?.definition?.name ?? 'Companion vignette order'}</h3>
-            {companionStabilityTranscripts.length > 0 ? (
-              <TranscriptList
-                transcripts={companionStabilityTranscripts}
-                onSelect={setSelectedTranscript}
-                groupByModel={false}
-                scenarioDimensions={companionScenarioDimensions}
-                onDecisionChange={handleDecisionChange}
-                updatingTranscriptIds={updatingTranscriptIds}
-                decisionColumnLabel={decisionColumnLabel}
-                decisionColumnTooltip={decisionColumnTooltip}
-                decisionDisplayMode={listDisplayMode}
-              />
-            ) : (
-              <div className="rounded-lg border border-gray-200 bg-white p-4 text-sm text-gray-500">
-                No transcripts matched this pattern for this vignette order.
-              </div>
-            )}
-          </section>
-        </div>
+        <PairedStabilityView
+          repeatPattern={repeatPattern}
+          primaryRunName={run?.definition?.name}
+          companionRunName={companionRun?.definition?.name}
+          primaryTranscripts={primaryStabilityTranscripts}
+          companionTranscripts={companionStabilityTranscripts}
+          primaryScenarioDimensions={scenarioDimensions}
+          companionScenarioDimensions={companionScenarioDimensions}
+          onSelectTranscript={setSelectedTranscript}
+          onDecisionChange={handleDecisionChange}
+          updatingTranscriptIds={updatingTranscriptIds}
+          decisionColumnLabel={decisionColumnLabel}
+          decisionColumnTooltip={decisionColumnTooltip}
+          decisionDisplayMode={listDisplayMode}
+        />
       ) : pairedConditionStateError ? (
         <ErrorMessage message={pairedConditionStateError.message} />
       ) : reportStateError ? (
